move postItem in post folder + remove dead code that uses thumbsUp
